Check file exists before trying to delete it.

+++ amsr2/get_jaxa_files.py
@@ -333,7 +333,9 @@ def get_hdfs_between_datetimes(start_dt, end_dt, output_dir=DATA_DIR, ftp_dir=FT
             # No files found, must be an empty bin.
             print("\tBin is empty, skipping.")
 
-            delete_file(last_file_of_prev_bin)
+            if last_file_of_prev_bin:
+                delete_file(last_file_of_prev_bin)
+                
             last_file_of_prev_bin = None
 
             continue
